+++ src/main/scala/org/broadinstitute/hail/driver/Main.scala
@@ -8,6 +8,7 @@ import org.apache.spark._
 import org.apache.spark.sql.SQLContext
 import org.broadinstitute.hail.FatalException
 import org.broadinstitute.hail.Utils._
+import org.broadinstitute.hail.io.gen.GenReport
 import org.broadinstitute.hail.methods.VCFReport
 import org.kohsuke.args4j.{CmdLineException, CmdLineParser, Option => Args4jOption}
 
@@ -111,6 +112,7 @@ object Main {
     }
 
     VCFReport.report()
+    GenReport.report()
 
     // Thread.sleep(60*60*1000)