+++ src/main/scala/org/broadinstitute/hail/driver/SplitMulti.scala
@@ -65,6 +65,7 @@ object SplitMulti extends Command {
     propagateGQ: Boolean,
     compress: Boolean,
     insertSplitAnnots: (Annotation, Int, Boolean) => Annotation): Iterator[(Variant, Annotation, Iterable[Genotype])] = {
+
     if (v.isBiallelic)
       return Iterator((v, insertSplitAnnots(va, 0, false), it))
 
@@ -76,49 +77,83 @@ object SplitMulti extends Command {
         (Variant(v.contig, newStart, newRef, newAlt), i + 1)
       }.toArray
 
-    val splitGenotypeBuilders = splitVariants.map { case (sv, _) => new GenotypeBuilder(sv) }
-    val splitGenotypeStreamBuilders = splitVariants.map { case (sv, _) => new GenotypeStreamBuilder(sv, compress) }
+    val splitGenotypeBuilders = splitVariants.map { case (sv, _) => new GenotypeBuilder(sv.nAlleles) }
+    val splitGenotypeStreamBuilders = splitVariants.map { case (sv, _) => new GenotypeStreamBuilder(sv.nAlleles, compress) }
 
     for (g <- it) {
 
       val gadsum = g.ad.map(gadx => (gadx, gadx.sum))
+      val isDosage = Genotype.flagHasDosage(g.flags)
 
       // svj corresponds to the ith allele of v
       for (((svj, i), j) <- splitVariants.iterator.zipWithIndex) {
         val gb = splitGenotypeBuilders(j)
-
         gb.clear()
-        g.gt.foreach { ggtx =>
-          val gtx = splitGT(ggtx, i)
-          gb.setGT(gtx)
 
-          val p = Genotype.gtPair(ggtx)
-          if (gtx != p.nNonRefAlleles)
-            gb.setFakeRef()
-        }
+        if (!isDosage) {
+          g.gt.foreach { ggtx =>
+            val gtx = splitGT(ggtx, i)
+            gb.setGT(gtx)
 
-        gadsum.foreach { case (gadx, sum) =>
-          // what bcftools does
-          // Array(gadx(0), gadx(i))
-          gb.setAD(Array(sum - gadx(i), gadx(i)))
-        }
+            val p = Genotype.gtPair(ggtx)
+            if (gtx != p.nNonRefAlleles)
+              gb.setFakeRef()
+          }
 
-        g.dp.foreach { dpx => gb.setDP(dpx) }
+          gadsum.foreach { case (gadx, sum) =>
+            // what bcftools does
+            // Array(gadx(0), gadx(i))
+            gb.setAD(Array(sum - gadx(i), gadx(i)))
+          }
 
-        if (propagateGQ)
-          g.gq.foreach { gqx => gb.setGQ(gqx) }
+          g.dp.foreach { dpx => gb.setDP(dpx) }
 
-        g.pl.foreach { gplx =>
-          val plx = gplx.iterator.zipWithIndex
-            .map { case (p, k) => (splitGT(k, i), p) }
-            .reduceByKeyToArray(3, Int.MaxValue)(_ min _)
-          gb.setPL(plx)
+          if (propagateGQ)
+            g.gq.foreach { gqx => gb.setGQ(gqx) }
 
-          if (!propagateGQ) {
-            val gq = Genotype.gqFromPL(plx)
-            gb.setGQ(gq)
+          g.pl.foreach { gplx =>
+            val plx = gplx.iterator.zipWithIndex
+              .map { case (p, k) => (splitGT(k, i), p) }
+              .reduceByKeyToArray(3, Int.MaxValue)(_ min _)
+            gb.setPL(plx)
+
+            if (!propagateGQ) {
+              val gq = Genotype.gqFromPL(plx)
+              gb.setGQ(gq)
+            }
+          }
+        } else {
+          val newdx = g.dosage.map { case gdx =>
+            val dx = gdx.iterator.zipWithIndex
+              .map { case (d, k) => (splitGT(k, i), d) }
+              .reduceByKeyToArray(3, 0.0)(_ + _)
+
+            gb.setDosage(dx)
+            dx
+          }
+
+          val newgt = newdx match {
+            case Some(x) => {
+              val maxDosage = x.max
+              if (maxDosage < 0.5 && x.count{d: Double => math.abs(d - maxDosage) <= 3e-4 } != 1) //first comparison for speed to not evaluate count if prob > 0.5
+                -1
+              else
+                x.indexOf(maxDosage)
+            }
+            case None => -1
+          }
+
+          if (newgt != -1)
+            gb.setGT(newgt)
+
+          //FIXME: don't know how to set the fakeref if the original genotype was -1
+          g.gt.foreach { gtx =>
+            val p = Genotype.gtPair(gtx)
+            if (newgt != p.nNonRefAlleles && newgt != -1)
+              gb.setFakeRef()
           }
         }
+
 
         splitGenotypeStreamBuilders(j).write(gb)
       }