+++ src/main/scala/org/broadinstitute/hail/methods/LoadVCF.scala
@@ -84,6 +84,7 @@ object VCFReport {
   }
 }
 
+
 object LoadVCF {
   def lineRef(s: String): String = {
     var i = 0