+++ src/test/scala/org/broadinstitute/hail/variant/GenotypeSuite.scala
@@ -16,11 +16,12 @@ object GenotypeSuite {
   def readWriteEqual(v: Variant, g: Genotype): Boolean = {
     ab.clear()
 
-    val gb = new GenotypeBuilder(v)
+    val gb = new GenotypeBuilder(v.nAlleles)
+
     gb.set(g)
     gb.write(ab)
-    val g2 = Genotype.read(v, new ByteIterator(ab.result()))
-    //println(s"g = $g, g2 = $g2")
+    val g2 = Genotype.read(v.nAlleles, new ByteIterator(ab.result()))
+
     g == g2
   }
 
@@ -89,6 +90,6 @@ class GenotypeSuite extends TestNGSuite {
     assert(D_==(Genotype(Some(1), Some(Array(16, 16)), Some(33), Some(99), Some(Array(100, 0, 100))).pAB().get, 1.0))
     assert(D_==(Genotype(Some(1), Some(Array(5, 8)), Some(13), Some(99), Some(Array(200, 0, 100))).pAB().get, 0.423950))
 
-    Spec.check()
+    Spec.check(size = 100, count = 100000)
   }
 }